+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+
+## [0.6] - 2020-05-06
 ### Added
 - Adding [zenbuild].
 - Add secondary run target to the make compiler.
@@ -79,7 +81,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 First major release.
 
 
-[Unreleased]: https://github.com/plandes/flex-compile/compare/v0.5...HEAD
+[Unreleased]: https://github.com/plandes/flex-compile/compare/v0.6...HEAD
+[0.6]: https://github.com/plandes/flex-compile/compare/v0.5...v0.6
 [0.5]: https://github.com/plandes/flex-compile/compare/v0.4...v0.5
 [0.4]: https://github.com/plandes/flex-compile/compare/v0.3...v0.4
 [0.3]: https://github.com/plandes/flex-compile/compare/v0.2...v0.3